requirements.txt
configparser>=3.5.0
discord.py[voice]>=0.16.2
colorlog>=3.1.0